fix pca 2

+++ novae/module/embedding.py
@@ -49,5 +49,5 @@ class GenesEmbedding(L.LightningModule):
         X = adata.X.toarray() if issparse(adata.X) else adata.X
 
         pca = PCA(n_components=self.embedding_size)
-        pca.fit(X)
+        pca.fit(X.astype(np.float32))
         self.embedding.weight.data = torch.tensor(pca.components_.T, device=self.device)
